requirements_backup.txt
backcall>=0.2.0
blis>=0.4.1
catalogue>=1.0.0
certifi>=2020.6.20
chardet>=3.0.4
click>=7.1.2
cloudpickle>=1.4.1
colorama>=0.4.3
cycler>=0.10.0
cymem>=2.0.3
cytoolz>=0.10.1
dask >=2.0, <3.0
decorator>=4.4.2
docopt>=0.6.2
idna>=2.9
imagecodecs>=2020.5.30
imageio>=2.8.0
ipython>=7.15.0
ipython-genutils>=0.2.0
jedi>=0.0
joblib>=0.0,<1.0
kiwisolver>=1.2.0
lightgbm>=2.3.1
matplotlib>=3.2.2
murmurhash>=1.0.2
natsort>=7.0.1
networkx>=2.4
nltk>=3.5
numpy >=1.0,<2.0
olefile>=0.46
pandas >=1.0,<2.0
pandas-flavor>=0.2.0
parso>=0.7.0
pickleshare>=0.7.5
Pillow>=7.1.2
plac>=1.1.3
preshed>=3.0.2
prompt-toolkit 
Pygments>=2.6.1
pyjanitor>=0.20.7
pyparsing>=2.4.7
Pyphen>=0.9.5
python-dateutil>=2.8.1
pytz>=2020.1
PyWavelets>=1.1.1
PyYAML>=5.3.1
regex>=2020.6.8
requests>=2.23.0
scikit-image>=0.17.2
scikit-learn>=0.0,<1.0
scipy>=1.4.1
shap>=0.35.0
six>=1.0,<2.0
spacy>=2.3.0
srsly>=1.0.2
textdistance>=4.2.0
textstat>=0.6.2
thinc>=7.4.1
threadpoolctl>=1.0,<3.0
tifffile>=2020.6.3
tika>=1.24
toolz>=0.10.0
tornado>=6.0.4
tqdm>=2.0
traitlets>=4.3.3
urllib3>=1.25.9
wasabi>=0.6.0
wcwidth>=0.0, <1.0
wincertstore>=0.2
xarray>=0.15.1
xgboost>=1.1.1
xlrd>=1.2.0
